requirement.txt
requests
pandas
sqlalchemy
psycopg2-binary
streamlit
python-dotenv